v1/requirements.txt
numpy==1.19.1
pydicom==2.1.2
pandas==1.0.5
scikit_image==0.17.2
matplotlib==3.2.1
tensorflow_gpu==2.2.0
opencv_python==4.2.0.34
tqdm==4.48.0
ipywidgets==7.5.1
scipy==1.4.1
fpdf==1.7.2
imutils==0.5.4
ipython==7.28.0
metrics==0.3.3
Pillow==8.4.0
scikit_learn==1.0.1
SimpleITK==2.1.1
skan==0.9
skimage==0.0
tensorflow==2.6.0
utils==1.0.1
